# openapi.yml
openapi: 3.1.0
info:
  title: AgoraLume API
  description: |-
    User + multi-AI group chat with modular personas. The backend owns the workspace and streams chat over SSE.

    POST creates (201), PATCH merges a partial body (200), DELETE returns 204. Failures are RFC 9457 problem documents; `type` is a stable `urn:agoralume:error:…`.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: 0.1.0
servers:
- url: /
  description: 'Same-origin: the bundled build, where the SPA and the API share a host'
- url: /api
  description: Behind an edge/dev proxy (VITE_API_PREFIX=/api)
- url: http://127.0.0.1:8080
  description: A local backend (pnpm dev:api)
paths:
...
  /v1beta/departments:
    get:
      tags:
      - departments
      summary: Every department in the workspace, across all organizations.
      operationId: list_depts
      responses:
        '200':
          description: Every department, across all organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - departments
      summary: Creates a department.
      operationId: create_dept
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
        required: true
      responses:
        '201':
          description: The created department
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1beta/departments/{departmentId}:
    get:
      tags:
      - departments
      summary: One department by id.
      operationId: get_dept
      parameters:
      - name: departmentId
        in: path
        description: The department to read
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The department
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown department
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - departments
      summary: Deletes a department.
      operationId: delete_dept
      parameters:
      - name: departmentId
        in: path
        description: The department to delete
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown department
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - departments
      summary: Merges a partial update onto a department.
      operationId: update_dept
      parameters:
      - name: departmentId
        in: path
        description: The department to update
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentPatch'
        required: true
      responses:
        '200':
          description: The updated department
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown department
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Malformed patch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1beta/groups:
    get:
      tags:
      - groups
      summary: Every chat group in the workspace.
      operationId: list_groups
      responses:
        '200':
          description: Every group
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - groups
      summary: Creates a group.
      operationId: create_group
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '201':
          description: The created group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1beta/groups/{groupId}:
    get:
      tags:
      - groups
      summary: One group by id.
      operationId: get_group
      parameters:
      - name: groupId
        in: path
        description: The group to read
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown group
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - groups
      summary: Deletes a group.
      operationId: delete_group
      parameters:
      - name: groupId
        in: path
        description: The group to delete
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown group
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - groups
      summary: Merges a partial update onto a group.
      operationId: update_group
      parameters:
      - name: groupId
        in: path
        description: The group to update
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupPatch'
        required: true
      responses:
        '200':
          description: The updated group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown group
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Malformed patch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1beta/personas/{personaId}/memories/{memoryId}:
    delete:
      tags:
      - personas
      summary: Deletes one of a persona's memories.
      operationId: delete_memory
      parameters:
      - name: personaId
        in: path
        description: The persona the memory belongs to
        required: true
        schema:
          type: string
      - name: memoryId
        in: path
        description: The memory to delete
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown memory for that persona
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    Department:
      type: object
      description: |-
        A sub-unit inside an organization; its variables sit between the org's and
        the persona's in the inheritance chain.
      required:
      - organizationId
      - name
      properties:
        blurb:
          type:
          - string
          - 'null'
        color:
          type:
          - string
          - 'null'
        id:
          type: string
        name:
          type: string
        organizationId:
          type: string
        variables:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
    DepartmentPatch:
      type: object
      description: 'A partial Department: only the properties present are merged. `id` is ignored.'
      properties:
        blurb:
          type:
          - string
          - 'null'
        color:
          type:
          - string
          - 'null'
        id:
          type: string
        name:
          type: string
        organizationId:
          type: string
        variables:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Access token from `POST /auth/login`. Opaque; does not survive a server restart. Waived when `GET /meta` reports `authRequired: false`.'
security:
- bearerAuth: []
...